Create README.md
                                                                                                                        CUSTOMER CARE REGISTRY

📃 PROJECT DESCRIPTION :
        
This Application has been developed to help the customer in processing their complaints.  The customers can raise the ticket with a detailed description of the issue.  An Agent will be assigned to the Customer to solve the problem.  Whenever the agent is assigned to a customer they will be notified with an email alert.  Customers can view the status of the ticket till the service is provided.
Admin : The main role and responsibility of the admin are to take care of the whole process.  Starting from Admin login followed by the agent creation and assigning the customer's complaints.  Finally, He will be able to track the work assigned to the agent and a notification will be sent to the customer.

User: They can register for an account.  After the login, they can create the complaint with a description of the problem they are facing.  Each user will be assigned with an agent.  They can view the status of their complaint.

TEAM DETAILS :
TEAM ID : PNT2022TMID13260
DOMAIN : CLOUD APP DEVELOPMENT

TEAM MEMBERS :
1.GOBINATH.M
2.KARUPPASAMY.N
3.MATHESH.S
4.SELVAKUMAR.S